GlobalMenu split from selection_menu.py
